+++ frontend/src/locales/en/channels.json
@@ -349,7 +349,10 @@
   "channels.dialogs.settings.overrides.action": "Overrides",
   "channels.dialogs.settings.overrides.title": "Override Settings",
   "channels.dialogs.settings.overrides.parameters.title": "Override Body",
-  "channels.dialogs.settings.overrides.parameters.description": "Provide JSON-formatted overrides to customize the request body sent through this channel. Use \"__AXONHUB_CLEAR__\" as the value to remove a parameter from the request body.",
+  "channels.dialogs.settings.overrides.parameters.description": "Configure request body parameters that will be applied to all requests sent through this channel. Use \"__AXONHUB_CLEAR__\" as the value to remove a parameter.",
+  "channels.dialogs.settings.overrides.parameters.key": "Parameter Key",
+  "channels.dialogs.settings.overrides.parameters.value": "Parameter Value",
+  "channels.dialogs.settings.overrides.parameters.addButton": "Add Parameter",
   "channels.dialogs.settings.overrides.headers.title": "Override Headers",
   "channels.dialogs.settings.overrides.headers.description": "Configure custom HTTP headers that will be applied to all requests sent through this channel.",
   "channels.dialogs.settings.overrides.headers.key": "Header Key",
