+++ frontend/src/locales/zh-CN/channels.json
@@ -352,7 +352,10 @@
   "channels.dialogs.settings.overrides.action": "覆盖设置",
   "channels.dialogs.settings.overrides.title": "覆盖配置",
   "channels.dialogs.settings.overrides.parameters.title": "覆盖请求体",
-  "channels.dialogs.settings.overrides.parameters.description": "提供 JSON 格式的覆盖配置，以定制通过此渠道发送的请求体。使用 \"__AXONHUB_CLEAR__\" 作为值可从请求体中移除参数。",
+  "channels.dialogs.settings.overrides.parameters.description": "配置请求体参数，会应用到通过此渠道发送的所有请求。使用 \"__AXONHUB_CLEAR__\" 作为值可移除参数。",
+  "channels.dialogs.settings.overrides.parameters.key": "参数 Key",
+  "channels.dialogs.settings.overrides.parameters.value": "参数 Value",
+  "channels.dialogs.settings.overrides.parameters.addButton": "新增参数",
   "channels.dialogs.settings.overrides.headers.title": "覆盖请求头",
   "channels.dialogs.settings.overrides.headers.description": "配置自定义 HTTP 请求头，会应用到通过此渠道发送的所有请求。",
   "channels.dialogs.settings.overrides.headers.key": "请求头 Key",
